\name{easyFMRI-package}
\alias{easyFMRI-package}
\alias{easyFMRI}
\docType{package}
\title{
Functions for FMRI Analysis
}
\description{
Functions for ROI segmentation in EPI sequences
}
\details{
\tabular{ll}{
Package: \tab easyFMRI\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-09-17\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Wonsang You

Maintainer: Wonsang You <you@ifn-magdeburg.de>
}
\references{
Wonsang You (2010) ROI Data Extraction from FMRI BOLD Signals of the Human Brain, Technical Reports of the Leibniz Institute for Neurobiology, TR 10016.
}
\keyword{ package }
\seealso{

}
\examples{
ROIdata<-orthoROIs()
}
